requirements.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.10.0.84
pillow==10.2.0
numpy==1.24.4
matplotlib==3.7.5